Split up the publishable module into two pieces [#98300946]
